Rename AssemblyMaker to ActlikeMaker
